finish part

+++ store/cassandra/models.go
@@ -3,7 +3,7 @@ package cassandra
 type OrderLine struct {
 	OrderLineId       int     `cql:"order_line_id"`
 	ItemId            int     `cql:"item_id"`
-	Amount            float64 `cql:"amount"`
+	Amount            float32 `cql:"amount"`
 	SupplyWarehouseId int     `cql:"supply_warehouse_id"`
 	Quantity          int     `cql:"quantity"`
 	MiscellaneousData string  `cql:"miscellaneous_data"`
